앱모듈 리펙토링
