+++ app/src/main/java/com/example/subway_alarm/module/SubwayAlarmApp.kt
@@ -1,0 +1,38 @@
+package com.example.subway_alarm.module
+
+import android.app.Application
+import android.content.Context
+import com.example.subway_alarm.data.api.ApiThread
+import com.example.subway_alarm.data.api.StationApi
+import com.example.subway_alarm.data.api.StationApiStorage
+import com.example.subway_alarm.data.repository.StationRepository
+import com.example.subway_alarm.data.repository.StationRepositoryImpl
+import com.example.subway_alarm.models.ViewModelImpl
+import org.koin.android.viewmodel.dsl.viewModel
+import org.koin.core.context.startKoin
+import org.koin.dsl.module
+
+class SubwayAlarmApp: Application() {
+    override fun onCreate() {
+        super.onCreate()
+
+        startKoin {
+            modules(appModule, appViewModule)
+        }
+    }
+    init{
+        instance = this
+    }
+
+    companion object {
+        lateinit var instance: SubwayAlarmApp
+
+        /**
+         * ApplicationContext가 필요할 때 호출하는 static 함수입니다.
+         */
+        fun ApplicationContext() : Context {
+            return instance.applicationContext
+        }
+    }
+
+}
